food_api/requirements.txt
fastapi[all]
beanie
motor
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
slowapi
pytest
pytest-asyncio
pytest-cov
httpx